Project1_3
Pretty much done except the server isn't responding to the client...?

+++ Request/Request.java
@@ -1,14 +1,17 @@
+/*
+ * This is the client's Request to the server.
+ */
 public class Request {
-    public int tml;            // Total message length (in bytes)
-    public int opCode;         // Desired operation code
-    public int op1;            // First operand
-    public int op2;            // Second operand
-    public int requestID;      // Differentiates requests
-    public int opNameLength;   // Length of operation name (in bytes)
+    public short tml;            // Total message length (in bytes)
+    public byte opCode;         // Desired operation code
+    public short op1;            // First operand
+    public short op2;            // Second operand
+    public byte requestID;      // Differentiates Requests
+    public byte opNameLength;   // Length of operation name (in bytes)
     public String opName;      // Operation name encoded with UTF-16
     
     
-    public Request(int tml, int opCode, int op1, int op2, int requestID, int opNameLength, String opName) {
+    public Request(short tml, byte opCode, short op1, short op2, byte requestID, byte opNameLength, String opName) {
         this.tml = tml;
         this.opCode = opCode;
         this.op1 = op1;
@@ -16,7 +19,7 @@ public class Request {
         this.requestID = requestID;
         this.opNameLength = opNameLength;
         this.opName = opName;
-    }
+        }
 
     public String toString() {
         final String EOLN = java.lang.System.getProperty("line.separator");
